Refactor to send provider data with payload
