+++ src/panel/components/DetailMenu.tsx
@@ -1,46 +1,12 @@
 import * as React from "react";
-import { Menu, Button, Icon, SemanticTRANSITIONS } from "semantic-ui-react";
+import { Table, Header, Container, Icon, Menu, Transition } from "semantic-ui-react";
+import { groupBy, defaultTo, map, keys, prop, sortBy } from "ramda";
+import { WebRequestParam } from "ski-providers";
 
-type State = {
-  copyText: string;
-};
+interface Props {
+  data: WebRequestParam[]
+}
 
-type Props = {
-  value: string;
-};
-
-export default class DetailMenu extends React.Component<Props, State> {
-  constructor(props: Props) {
-    super(props);
-    this.state = { copyText: 'Copy' }
-  }
-
-  showCopiedLabel() {
-    this.setState({ copyText: 'Copied!' });
-    setTimeout(
-      () => this.showCopyLabel(),
-      2000
-    );
-  }
-
-  showCopyLabel(){
-    this.setState({ copyText: 'Copy' });
-  }
-
-  onCopy() {
-    this.showCopiedLabel();
-  }
-
-  render() {
-    return (
-      <Menu compact size="mini" secondary>
-        <Menu.Item>
-          <Button basic onClick={this.onCopy.bind(this)} data-clipboard-text={this.props.value} className="clipBoard">
-            <Icon name="copy" />
-            {this.state.copyText}
-          </Button>
-        </Menu.Item>
-      </Menu>
-    );
-  }
+export const Payload = (props: Props): JSX.Element => {
+  return <div></div>
 }
